YAML/requirements.txt
# YAML Environment Requirements
# Install with: pip install -r requirements.txt

PyYAML>=6.0
